\name{Kendall-package}
\alias{Kendall-package}
\alias{Kendall}
\docType{package}
\title{
 Kendall Tau Implementation
}
\description{
  Implementation En R et Rcpp de deux algos calculant le Tau de Kendall
}
\author{
ASTY Guillaume, MASSEAU Armand, MORHAG Idris

Maintainer: ASTY Guillaume, MASSEAU Armand, MORHAG Idris
}
\keyword{ package }