% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{naive_tau_kendall_cpp}
\alias{naive_tau_kendall_cpp}
\title{Calcul du tau de Kendall avec une approche naive}
\usage{
naive_tau_kendall_cpp(x, y)
}
\arguments{
\item{x}{NumericVector, premier vecteur de donnees}

\item{y}{NumericVector, second vecteur de donnees}
}
\value{
Une valeur numerique representant le tau de Kendall
}
\description{
Calcul du tau de Kendall avec une approche naive
}
