% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NDtau_kendall_cpp}
\alias{NDtau_kendall_cpp}
\title{Calcul du tau de Kendall optimise}
\usage{
NDtau_kendall_cpp(X, Y)
}
\arguments{
\item{X}{NumericVector, premier vecteur de donnees}

\item{Y}{NumericVector, second vecteur de donnees}
}
\value{
Une valeur numerique representant le tau de Kendall
}
\description{
Calcul du tau de Kendall optimise
}
